Composite Gates

This module is used for gates that can be composed of more fundamental gate operations. For simplicity, we refer to these as Composite Gates, and represent them as their numerical matrix values.

CompositeGates.GateSymbols

As in DefaultGates, we store symbols for the gates available in CompositeGates, and generator functions for parameteric variants of these gates. This allows us to track the mappings as to be used by the gate type hierarchy and architecture.

Two Qubit Gates: static

QXZoo.CompositeGates.swap — Method

swap(q_target1::Int, q_target2::Int)

Generate a SWAP gate equivalent circuit, composed of fundamental gate operations (CX)

Examples

QXZoo.CompositeGates.iswap — Method

iswap(q_target1::Int, q_target2::Int)

Generate a iSWAP gate equivalent circuit, generatable using exp(1im*(pi/4)*(kron(GateMap.x(),GateMap.x()) + kron(GateMap.y(),GateMap.y())))

Examples

Two Qubit Gates: rotation

QXZoo.CompositeGates.xx — Method

xx(q_target1::Int, q_target2::Int, theta::Real)

Generate a XX parametric gate-equivalent circuit, composed of fundamental gate operations (H, CX, Rz)

Examples

QXZoo.CompositeGates.yy — Method

yy(q_target1::Int, q_target2::Int, theta::Real)

Generate a YY parametric gate, which can be constructed from fundamental 1 & 2 qubit operations

Examples

QXZoo.CompositeGates.zz — Method

zz(q_target1::Int, q_target2::Int, theta::Real)

Generate a ZZ parametric gate, which can be constructed from fundamental 1 & 2 qubit operations

Examples