Gate Operations

Gate type-system hierarchy

Abstraction of types for representing gates.

QXZoo.GateOps.AGate — Type

Abstract Gate type

QXZoo.GateOps.AGateSymbol — Type

Abstract Gate Label, for unique gates

QXZoo.GateOps.AGateCall — Type

Abstract Gate Call, for tracking Gate labels applied to specific qubits

GateSymbol and GateCall

Specific gate instance in a circuit, applied to a given (set of) qubit(s).

QXZoo.GateOps.GateSymbol — Type

Gate label. Tracks the gate symbol (:x,:y,:z, etc)

QXZoo.GateOps.GateSymbolP — Type

Gate label with params. Tracks the gate symbol (:x,:y,:z, etc)

QXZoo.GateOps.GateCall1 — Type

Single qubit Gate call. Has a GateSymbol, target qubit index

QXZoo.GateOps.GateCall2 — Type

Two qubit Gate call. Has a GateSymbol, control and target qubit indices, and base gate.

QXZoo.GateOps.GateCallN — Type

n-qubit Gate call. Has a GateSymbol, control vector indices, target qubit index, and base gate.

QXZoo.GateOps.Gate — Type

Wrap of Matrix to fit type system hierarchy